// front/src/models/models.tsx
export class userInfo {
        firstName: string;
        lastName: string;
        image: string[];
        href: string;
        aboutMe: string;
        dob: string;
        email: string;
        phoneNumber: string;
        school: string;
        year: string;
        resume: any;
        diploma: string;
        desiredWork: string;
}